+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -1,7 +1,9 @@
 package com.techcourse.controller;
 
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.context.stereotype.Controller;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
@@ -9,11 +11,12 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginController implements Controller {
+@Controller
+public class LoginController{
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    @Override
+    @RequestMapping(value = "/login", method = RequestMethod.POST)
     public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
         if (UserSession.isLoggedIn(req.getSession())) {
             return ModelAndView.ofJspView("redirect:/index.jsp");
