+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,13 +1,16 @@
 package com.techcourse.controller;
 
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public class RegisterViewController implements Controller {
+@Controller
+public class RegisterViewController {
 
-    @Override
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
     public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
         return ModelAndView.ofJspView("/register.jsp");
     }
