+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ControllerAdapter.java
@@ -1,0 +1,22 @@
+//package com.interface21.webmvc.servlet.mvc.tobe;
+//
+//import com.interface21.webmvc.servlet.ModelAndView;
+//import com.interface21.webmvc.servlet.mvc.asis.Controller;
+//import jakarta.servlet.http.HttpServletRequest;
+//import jakarta.servlet.http.HttpServletResponse;
+//
+//public class ControllerAdapter implements HandlerAdapter {
+//    @Override
+//    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+//        try {
+//            return ((Controller) handler).execute(request, response);
+//        } catch (Exception e) {
+//            throw new RuntimeException(e);
+//        }
+//    }
+//
+//    @Override
+//    public boolean support(Object handler) {
+//        return handler instanceof Controller;
+//    }
+//}
